refactor: seperate component
