+++ src/lib/components/common/Button/index.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import BasicButton from './style';
+
+const Button = ({ className, icon, disabled, onClick }) => {
+  return (
+    <BasicButton disabled={disabled} className={className} onClick={onClick}>
+      {icon}
+    </BasicButton>
+  );
+};
+
+export default Button;
