+++ src/lib/constants/constants.js
@@ -1,0 +1,48 @@
+const items = [
+  {
+    id: 1,
+    content: '1',
+  },
+  {
+    id: 2,
+    content: '2',
+  },
+  {
+    id: 3,
+    content: '3',
+  },
+  {
+    id: 4,
+    content: '4',
+  },
+  {
+    id: 5,
+    content: '5',
+  },
+  {
+    id: 6,
+    content: '6',
+  },
+  {
+    id: 7,
+    content: '7',
+  },
+  {
+    id: 8,
+    content: '8',
+  },
+  {
+    id: 9,
+    content: '9',
+  },
+  {
+    id: 10,
+    content: '10',
+  },
+  {
+    id: 11,
+    content: '11',
+  },
+];
+
+export default items;
